Strip out App code now that Calendar is in use
