+++ packages/cal-ui/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@insightorchards/calendar-ui",
-  "version": "0.0.40",
+  "version": "0.0.41",
   "description": "A calendar UI package",
   "author": "Insight Orchards",
   "main": "dist/index.js",
