+++ ui/package.json
@@ -15,7 +15,7 @@
     "@fullcalendar/list": "^5.11.3",
     "@fullcalendar/react": "^5.11.2",
     "@fullcalendar/timegrid": "^5.11.3",
-    "@insightorchards/calendar-ui": "^0.0.40",
+    "@insightorchards/calendar-ui": "^0.0.41",
     "@testing-library/jest-dom": "^5.16.5",
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
