+++ ui/typing.d.ts
@@ -1,2 +1,3 @@
 declare module "*.module.css";
 declare module "react-big-calendar";
+declare module "@insightorchards/calendar-ui";
